Fixed performance issue.

+++ core/src/main/java/tmb/randy/tmbgriefergames/core/events/MoneyBalanceChangedEvent.java
@@ -1,22 +1,21 @@
 package tmb.randy.tmbgriefergames.core.events;
 
 import net.labymod.api.event.Event;
-import java.math.BigDecimal;
 
 public class MoneyBalanceChangedEvent implements Event {
-    BigDecimal oldBalance;
-    BigDecimal newBalance;
+    double oldBalance;
+    double newBalance;
 
-    public MoneyBalanceChangedEvent(BigDecimal oldBalance, BigDecimal newBalance) {
+    public MoneyBalanceChangedEvent(double oldBalance, double newBalance) {
         this.newBalance = newBalance;
         this.oldBalance = oldBalance;
     }
 
-    public BigDecimal getOldBalance() {
+    public double getOldBalance() {
         return oldBalance;
     }
 
-    public BigDecimal getNewBalance() {
+    public double getNewBalance() {
         return newBalance;
     }
 }
